requirements.txt
Django==3.0.8
djangorestframework==3.11.0
flake8==3.8.3
django-cors-headers==3.4.0
easy-thumbnails==2.7
pillow==7.2.0